Make PHPMD happier
