+++ src/DOMNodeList.php
@@ -36,42 +36,3 @@ class DOMNodeList extends XMLDecorator implements \Iterator {
 		return new DOMNodeListPairsIterator($this);
 	}
 }
-
-/**
- * Iteratres over a LyteDOMNodeList as key value pairs
- */
-class DOMNodeListPairsIterator implements \Iterator {
-	private $_list;
-	private $_pos = 0;
-
-	public function __construct(DOMNodeList &$list) {
-		$this->_list =& $list;
-		$this->_seek();
-	}
-
-	public function current() {
-		return $this->_list->item($this->_pos)->nodeValue;
-	}
-	private function _seek() {
-		// iterate past non LyteDOMElement nodes
-		while ($this->valid() && !($this->_list->item($this->_pos) instanceof DOMElement)) {
-			$this->_pos++;
-		}
-	}
-	public function next() { 
-		$this->_pos++;
-		$this->_seek();
-	}
-	public function key() { 
-		return $this->_list->item($this->_pos)->nodeName; 
-	}
-	public function valid() {
-		return 
-			$this->_pos < $this->_list->length 
-			&& $this->_pos >= 0;
-	}
-	public function rewind() {
-		$this->_pos = 0;
-		$this->_seek();
-	}
-}
